requirements.txt
numpy
streamlit
streamlit-option-menu
Pillow
